Moving methods common to the View Models to a Helper class.
